Lets actually use the review data from shopify where possible.
Also moves the reviews API call down to the reviews section layer, the only place where that data is needed.
